requirements.txt
ultralytics==8.4.13
torch_pruning
pyyaml
torch